Move tree widget into UI package
