+++ FyneApp.toml
@@ -4,8 +4,8 @@ Website = "https://github.com/ErikKalkoken/janice"
   Icon = "icon.png"
   Name = "Janice"
   ID = "io.github.erikkalkoken.janice"
-  Version = "0.4.2"
-  Build = 2
+  Version = "0.5.0"
+  Build = 1
 
 [Release]
   BuildName = "janice"
